AllGreenTest

+++ src/facades/PersonFacadeDB.java
@@ -7,9 +7,12 @@ package facades;
 
 import com.google.gson.Gson;
 import exceptions.NotFoundException;
+import java.util.Collections;
+import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
+import javax.persistence.Query;
 import model.Person;
 
 /**
@@ -21,12 +24,12 @@ public class PersonFacadeDB implements IPersonFacade {
     private static PersonFacadeDB instance = new PersonFacadeDB();
 
     public static PersonFacadeDB getFacade(boolean b) {
-        
-    if(true){
-      instance = new PersonFacadeDB();
-    }
-    return instance;
-  
+
+        if (true) {
+            instance = new PersonFacadeDB();
+        }
+        return instance;
+
     }
     EntityManagerFactory emf = Persistence.createEntityManagerFactory("RestCRUDPU");
     EntityManager em = emf.createEntityManager();
@@ -35,11 +38,10 @@ public class PersonFacadeDB implements IPersonFacade {
     @Override
     public Person addPerson(String json) {
         Person p = gson.fromJson(json, Person.class);
-        
-        System.out.println(p.getPhone());
         em.getTransaction().begin();
         em.persist(p);
         em.getTransaction().commit();
+
         return p;
 
     }
@@ -47,23 +49,37 @@ public class PersonFacadeDB implements IPersonFacade {
     @Override
     public Person deletePerson(int id) throws NotFoundException {
         Person p = em.find(Person.class, id);
+        if (p == null) {
+            throw new NotFoundException("No person exists for the given id");
+        }
+
         em.getTransaction().begin();
         em.remove(p);
         em.getTransaction().commit();
+
         return p;
     }
 
     @Override
     public String getPerson(int id) throws NotFoundException {
         Person p = em.find(Person.class, id);
-        return p.toString();
+         if (p == null) {
+            throw new NotFoundException("No person exists for the given id");
+        }
+        return gson.toJson(p);
     }
 
     @Override
     public String getPersons() {
-        String p = em.createQuery("SELECT p FROM Person p").toString();
+        Query p = em.createQuery("SELECT p FROM Person p");
 
-        return p;
+        List pe = p.getResultList();
+
+        System.out.println(pe);
+
+        Collections.reverse(pe);
+
+        return gson.toJson(pe);
     }
 
     @Override
@@ -73,7 +89,7 @@ public class PersonFacadeDB implements IPersonFacade {
         if (oldValue == null) {
             throw new NotFoundException("No person exists for the given id");
         }
-        
+
         em.persist(p);
         return oldValue;
     }
